+++ poc/classes/ContextNamespaceIdentifier.py
@@ -1,6 +1,6 @@
 from poc.classes.AuxISourceAnalyser import AuxISourceAnalyser
 from poc.classes.AuxInterpretation import AuxInterpretation
-from poc.classes.AuxOutlines import AuxOutlines
+from poc.classes.AuxContext import AuxContext
 from poc.classes.NamespaceIdentifier import NamespaceIdentifier
 from poc.classes.AuxSymbolTable import AuxSymbolTable
 
@@ -10,18 +10,15 @@ class ContextNamespaceIdentifier:
     def dispatch(i: AuxISourceAnalyser, parsing_info: AuxInterpretation):
         namespace_info = NamespaceIdentifier(i.parse_list, parsing_info)
         # NamespaceIdentifer can occur in the following contexts:
-        if i.context.is_parsing_context([AuxOutlines.root]):
+        if i.context.is_parsing_context([AuxContext.root]):
             # tag the namespace to the theory node
             i.theory_node.namespace = namespace_info.id
-        elif i.context.is_parsing_context([AuxOutlines.root, AuxOutlines.block, AuxOutlines.uses]):
+        elif i.context.is_parsing_context([AuxContext.root, AuxContext.block, AuxContext.uses]):
             # as a name of namespaces used in the current namespace
-            # In this case, remember the node of the used namespace added to the symbol table
-            # because we will need it again when handling the NamespaceModifier rule (see last_element_of_list_or_tuple)
-            i.working_stack.append(
-                AuxSymbolTable.add_usage_to_theory(i.theory_node, namespace_info))
-
+            # Remember the used namespace because we will need it in case ContextWildcardTheoryNamespace is called
+            i.push_node(AuxSymbolTable.add_usage_to_theory(i.theory_node, namespace_info))
         else:
-            if i.debug:
+            if i.verbose:
                 print(
                     "########### Unhandled context in ContextNamespaceIdentifier.dispatch " + str(
                         i.context.get_context()) + " " + str(namespace_info))