+++ poc/classes/ContextPredicateHeader.py
@@ -1,6 +1,6 @@
 from poc.classes.AuxISourceAnalyser import AuxISourceAnalyser
 from poc.classes.AuxInterpretation import AuxInterpretation
-from poc.classes.AuxOutlines import AuxOutlines
+from poc.classes.AuxContext import AuxContext
 from poc.classes.ContextPredicateDeclaration import ContextPredicateDeclaration
 
 
@@ -8,17 +8,17 @@ class ContextPredicateHeader:
     @staticmethod
     def dispatch(i: AuxISourceAnalyser, parsing_info: AuxInterpretation):
         # a predicate header can occur in the following contexts:
-        if i.context.is_parsing_context([AuxOutlines.theory, AuxOutlines.block]):
+        if i.context.is_parsing_context([AuxContext.theory, AuxContext.block]):
             # inside a theory as a global predicate definition
-            ContextPredicateDeclaration.start(i,parsing_info)
+            ContextPredicateDeclaration.start(i, parsing_info)
             return
-        elif i.context.is_parsing_context([AuxOutlines.optionalProperty]) or \
-                i.context.is_parsing_context([AuxOutlines.mandatoryProperty]):
-            # inside a class, or functional term, a predicate instance
-            ContextPredicateDeclaration.start(i,parsing_info)
+        elif i.context.is_parsing_context([AuxContext.optionalProperty]) or \
+                i.context.is_parsing_context([AuxContext.mandatoryProperty]):
+            # inside a class, or functional term as a predicate property
+            ContextPredicateDeclaration.start(i, parsing_info)
             return
         else:
-            if i.debug:
+            if i.verbose:
                 print("########### Unhandled context in ContextPredicateHeader.dispatch" + str(
                     i.context.get_context()) + " " + str(parsing_info))
         i.parse_list.append(parsing_info)