+++ poc/fplmessage.py
@@ -2,11 +2,6 @@ import re
 
 
 class FplParserError(Exception):
-    __line = 0
-    __col = 0
-    __msg = ""
-    __type = None
-    mainType = None
 
     def __init__(self, inner_exception, error_context):
         self.error_context = error_context
@@ -16,13 +11,17 @@ class FplParserError(Exception):
             self.__line = result.group(1)
             self.__col = result.group(2)
             self.__msg = str(inner_exception)
+        else:
+            self.__line = 0
+            self.__col = 0
+            self.__msg = ""
 
         self.inner_exception = inner_exception
         self.__type = str(type(self).__name__)
         self.mainType = "E"
 
     def __str__(self):
-        msg = self.__type
+        msg = self.__type + " "
         pos = str(self.error_context).find("\n")
         if pos > -1:
             msg += self.error_context[:pos]
@@ -46,18 +45,14 @@ class FplParserError(Exception):
 
 
 class FplInterpreterMessage(Exception):
-    __msg = None
-    __line = None
-    __col = None
-    __type = None
-    mainType = None
 
-    def __init__(self, msg: str, line: int, col: int):
+    def __init__(self, msg: str, line: int, col: int, file: str):
         self.__msg = msg
         self.__line = line
         self.__col = col
         self.__type = str(type(self).__name__)
         self.mainType = "E"  # Error per default
+        self.file = file
 
     def __str__(self):
         return self.__type + ":" + str(self.__line) + ":" + str(self.__col) + ": " + self.__msg
@@ -76,5 +71,3 @@ class FplInterpreterMessage(Exception):
 
     def to_tuple(self):
         return self.__type, self.__msg, self.__line, self.__col
-
-