+++ poc/CHANGES.md
@@ -36,3 +36,8 @@ Initial interpreter
 * Bugfix instance namespaces for variables of auxiliary classes
 * Separation of syntax transforming (minify/prettify) and syntax analysis
 * Redesign and simplification of the FPLSourceAnalyser class while moving context handling to individual classes named Contex<Something>
+## 1.2.1
+* Class AuxOutline was removed. It mixed up the symbols used for the outline of the symbol table (now in the class AuxSymbolTable) with the symbols used for contexts (now in the class AuxContext). 
+* Symbol table functionality variable declarations (cont.)
+* New tests added
+* paremetrised tests